corrected hw03
