+++ hwDinaScanner/src/de/telran/application/MainScanner.java
@@ -1,21 +1,20 @@
 package de.telran.application;
-import de.telran.data.ScannerProject;
 
-import java.util.Scanner;
-
-import static de.telran.data.ScannerProject.*;
+import static de.telran.data.ScannerProjectArray.fillArray;
+import static de.telran.data.ScannerProjectArray.isDeleteNumberInArray;
+import static de.telran.data.UserInteraction.*;
 
 public class MainScanner {
-    static Scanner scanner = new Scanner(System.in);
+
     public static void main(String[] args) {
-        System.out.println("Enter size of array, please");
-        int arraySize = scanner.nextInt();
+
+        int arraySize = getArraySize();
         int[] myArray = new int[arraySize];
         fillArray(myArray);
-        System.out.println("Enter number to delete");
-        int deleteNumber = scanner.nextInt();
+
+        int deleteNumber = getDeletedNumber();
         boolean checkNumber = isDeleteNumberInArray(myArray, deleteNumber);
         printResult(myArray, checkNumber, deleteNumber);
-        scanner.close();
+        closeScanner();
     }
 }
